+++ src/main/java/org/realityforge/gwt/websockets/client/html5/Html5WebSocket.java
@@ -56,6 +56,12 @@ public class Html5WebSocket
   }
 
   @Override
+  public String getProtocol()
+  {
+    return _webSocket.getProtocol();
+  }
+
+  @Override
   public final void send( @Nonnull String data )
   {
     checkConnected();
@@ -120,6 +126,10 @@ public class Html5WebSocket
 
     native void send( String data ) /*-{
       this.send( data );
+    }-*/;
+
+    native String getProtocol()  /*-{
+        return this.protocol;
     }-*/;
   }
 }